CA-781 Renamed state/index to AppStateProvider
